chore: separate console to an individual package
