+++ packages/bottender-console/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@bottender/console",
+  "version": "1.5.4",
+  "license": "MIT",
+  "homepage": "https://bottender.js.org/",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Yoctol/bottender.git"
+  },
+  "main": "dist/index.js",
+  "files": [
+    "dist"
+  ],
+  "types": "dist/index.d.ts",
+  "dependencies": {
+    "@bottender/core": "file:../bottender-core",
+    "delay": "^5.0.0",
+    "facebook-batch": "^1.1.0",
+    "messaging-api-common": "^1.0.4",
+    "messaging-api-messenger": "^1.1.0",
+    "shortid": "^2.2.16",
+    "ts-invariant": "^0.9.3",
+    "type-fest": "^2.5.0",
+    "warning": "^4.0.3"
+  }
+}
