+++ packages/bottender-console/src/__tests__/ConsoleBot.spec.ts
