+++ packages/bottender-console/src/index.ts
@@ -1,0 +1,4 @@
+export { default as ConsoleBot } from './ConsoleBot';
+export { default as ConsoleConnector } from './ConsoleConnector';
+export { default as ConsoleContext } from './ConsoleContext';
+export { default as ConsoleEvent } from './ConsoleEvent';
